+++ directory_components/middleware.py
@@ -102,7 +102,7 @@ class CountryMiddleware(MiddlewareMixin):
                 value=request.COUNTRY_CODE,
                 max_age=settings.LANGUAGE_COOKIE_AGE,
                 path=settings.LANGUAGE_COOKIE_PATH,
-                domain=settings.LANGUAGE_COOKIE_DOMAIN
+                domain=settings.LANGUAGE_COOKIE_DOMAIN,
             )
         return response
 
@@ -127,7 +127,9 @@ class PersistLocaleMiddleware(MiddlewareMixin):
             value=translation.get_language(),
             max_age=settings.LANGUAGE_COOKIE_AGE,
             path=settings.LANGUAGE_COOKIE_PATH,
-            domain=settings.LANGUAGE_COOKIE_DOMAIN
+            domain=settings.LANGUAGE_COOKIE_DOMAIN,
+            secure=True,
+            httponly=True,
         )
         return response